+++ scripts/ExtensionAttributes/FusionSerialNumbers.sh
@@ -4,7 +4,7 @@
 # Function: Check if VMware Fusion is installed, and gather the serial numbers associated with it
 # Requirements: None
 #
-# Copyright 2018, Route 1337, LLC, All Rights Reserved.
+# Copyright 2020, Route 1337, LLC, All Rights Reserved.
 #
 # Maintainers:
 # - Matthew Ahrenstein: matthew@route1337.com